requirements.txt
# 核心数据科学库
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# 机器学习框架
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.1.0
catboost==1.2.2
torch==2.1.0
torchvision==0.16.0
deap==1.4.1  # 遗传编程

# 可视化库
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
graphviz==0.20.1
networkx==3.1
# pygraphviz==1.11  # 网络图可视化

# 数据处理和实验
imbalanced-learn==0.11.0
optuna==3.3.0  # 超参数优化
mlflow==2.8.1  # 实验跟踪
wandb==0.15.12  # 实验跟踪（可选）
joblib==1.3.2

# 项目工具
jupyter==1.0.0
ipython==8.14.0
tqdm==4.65.0
colorama==0.4.6
rich==13.5.2  # 终端美化

# 文档和报告
pypandoc==1.11
jupyter-book==0.15.1
mkdocs==1.5.3
mkdocs-material==9.4.2

# 测试和代码质量
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# 配置文件管理
pyyaml==6.0.1
toml==0.10.2

# 其他工具
pathlib==1.0.1  # Python 3.4+内置，但为兼容性保留
python-dotenv==1.0.0
openpyxl==3.1.2  # Excel文件处理